use stricted threshold in full tortoise
