+++ mesh/meshdb.go
@@ -289,14 +289,14 @@ func (m *DB) ContextualValidity(id types.BlockID) (bool, error) {
 }
 
 // SaveContextualValidity persists opinion on block to the database.
-func (m *DB) SaveContextualValidity(id types.BlockID, _ types.LayerID, valid bool) error {
+func (m *DB) SaveContextualValidity(id types.BlockID, lid types.LayerID, valid bool) error {
 	var v []byte
 	if valid {
 		v = constTrue
 	} else {
 		v = constFalse
 	}
-	m.With().Debug("save block contextual validity", id, log.Bool("validity", valid))
+	m.With().Debug("save block contextual validity", id, lid, log.Bool("validity", valid))
 
 	if err := m.contextualValidity.Put(id.Bytes(), v); err != nil {
 		return fmt.Errorf("put into DB: %w", err)
@@ -993,6 +993,7 @@ func (m *DB) LayerContextuallyValidBlocks(ctx context.Context, layer types.Layer
 
 	cvErrors := make(map[string][]types.BlockID)
 	cvErrorCount := 0
+
 	for _, b := range blockIds {
 		valid, err := m.ContextualValidity(b)
 		if err != nil {
